[NFC][TransformStrategies] Abstract out a GEMM like strategy from MatmulTensorCoreStrategy
In preparation for other GEMM-like transform dialect strategies, this
makes the strategy building helpers for the existing matmul strategy
available for use elsewhere.
