+++ compiler/src/iree/compiler/Codegen/TransformDialectStrategies/GPU/Common.h
@@ -7,6 +7,7 @@
 #ifndef IREE_COMPILER_CODEGEN_TRANSFORM_DIALECT_STRATEGIES_GPU_COMMON_H_
 #define IREE_COMPILER_CODEGEN_TRANSFORM_DIALECT_STRATEGIES_GPU_COMMON_H_
 
+#include "iree/compiler/Codegen/TransformDialectStrategies/GPU/AbstractGemmLikeStrategy.h"
 #include "iree/compiler/Codegen/TransformDialectStrategies/GPU/AbstractReductionStrategy.h"
 #include "llvm/ADT/StringRef.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
@@ -85,6 +86,67 @@ void build1DSplittingStrategyWithOptionalThreadMapping(
     ImplicitLocOpBuilder& b, Value isolatedParentOpH, Value opH, int64_t rank,
     int64_t mostMinorDim, SmallVector<int64_t> opSizes, int64_t numThreads,
     Attribute mappingAttr = Attribute(), int64_t maxVectorSize = 4);
+
+/// Build the transform IR to pad a matmul op `matmulOpH`.
+Value buildPadMatmul(ImplicitLocOpBuilder &b, Value matmulOpH,
+                     const AbstractGemmLikeStrategy &strategy);
+
+/// Build transform IR to hoist the padded output operand of a padded matmul.
+/// Additionally, this attempts to fold the padding into the producing fill, if
+/// available.
+// TODO: Generalize, this is not specific to a matmul.
+// TODO: Better API
+Value buildHoistOutputPaddingOp(ImplicitLocOpBuilder &b, Value variantH,
+                                Value paddedMatmulOpH,
+                                int64_t numLoopsToHoist = 1);
+
+/// Helper function to distribute one pad or copy operation.
+/// Note: When `foldIfBranch` is true, one must later perform masked
+/// vectorization of the result.
+/// This amounts to injecting knowledge about future transformations without
+/// adding leaky semantics.
+Value buildDistributeOnePadOrCopy(ImplicitLocOpBuilder &b, Value variantH,
+                                  Value copyOpH, ArrayRef<int64_t> numThreads,
+                                  ArrayRef<Attribute> threadDimMapping,
+                                  bool foldIfBranch = false);
+
+/// Distribute the explicit copies involved in a matmul operation
+/// `paddedMatmulOpH`.
+std::tuple<Value, Value, Value> buildDistributeCopies(
+    ImplicitLocOpBuilder &b, Value variantH, Value paddedMatmulOpH,
+    const AbstractGemmLikeStrategy &strategy);
+
+/// Specific pattern to perform masked vectorization of copies give as
+/// parameters, cleanup and vectorize the rest.
+void buildMatmulVectorization(ImplicitLocOpBuilder &b, Value variantH,
+                              Value lhsCopyOpH, Value rhsCopyOpH,
+                              Value copyBackOpH,
+                              const AbstractGemmLikeStrategy &strategy);
+
+/// Build the transform IR to perform conversion to tensor core operations.
+/// This is currently subject to phase orderings as follows:
+///   - Vector transfer_read and transfer_write patterns have different subview
+///     folding behavior, force a fold_memref_aliases on them to enable
+///     redundant vector transfer hoisting.
+///   - Unfortunately, fold_memref_aliases breaks vector_to_mma conversion
+///     across scf.for after unrolling due to insert_strided_slice /
+///     extract_strided_slice across iter_args boundaries.
+///   - Hoist redundant vector transfers to allow conversion to tensor core to
+///     proceed. We really don't want to do this after bufferization but we need
+///     to atm.
+Value buildConvertToTensorCoreOp(ImplicitLocOpBuilder &b, Value funcH,
+                                 const AbstractGemmLikeStrategy &strategy);
+
+void buildMultiBuffering(ImplicitLocOpBuilder &b, Value funcH,
+                         const AbstractGemmLikeStrategy &strategy);
+
+Value buildConvertToAsyncCopies(ImplicitLocOpBuilder &b, Value funcH,
+                                const AbstractGemmLikeStrategy &strategy);
+
+void buildPipelineSharedMemoryCopies(ImplicitLocOpBuilder &b, Value funcH,
+                                     const AbstractGemmLikeStrategy &strategy);
+
+Value buildBufferize(ImplicitLocOpBuilder &b, Value variantH);
 
 //===----------------------------------------------------------------------===//
 // Higher-level problem-specific strategy creation APIs, these should favor
