+++ compiler/src/iree/compiler/Codegen/TransformDialectStrategies/GPU/MatmulTensorCoreStrategy.h
@@ -8,6 +8,7 @@
 #define IREE_COMPILER_CODEGEN_TRANSFORM_DIALECT_STRATEGIES_GPU_TENSOR_CORE_MATMUL_STRATEGY_H_
 
 #include "iree-dialects/Transforms/TransformMatchers.h"
+#include "iree/compiler/Codegen/TransformDialectStrategies/GPU/AbstractGemmLikeStrategy.h"
 #include "mlir/Dialect/GPU/IR/GPUDialect.h"
 
 namespace llvm {
@@ -20,126 +21,49 @@ namespace gpu {
 
 struct GPUModel;
 
-/// Base quantities generally useful for all GPU strategies.
-// TODO: refactor into a common place.
-struct StrategyBase {
-  StrategyBase(MLIRContext *ctx) : ctx(ctx) {}
+using iree_compiler::gpu::AbstractGemmLikeStrategy;
 
-  /// Constructor quantities.
-  MLIRContext *ctx;
-
-  Attribute blockX() const {
-    return mlir::gpu::GPUBlockMappingAttr::get(ctx, mlir::gpu::Blocks::DimX);
-  }
-  Attribute blockY() const {
-    return mlir::gpu::GPUBlockMappingAttr::get(ctx, mlir::gpu::Blocks::DimY);
-  }
-  Attribute blockZ() const {
-    return mlir::gpu::GPUBlockMappingAttr::get(ctx, mlir::gpu::Blocks::DimZ);
-  }
-  Attribute threadX() const {
-    return mlir::gpu::GPUThreadMappingAttr::get(ctx, mlir::gpu::Threads::DimX);
-  }
-  Attribute threadY() const {
-    return mlir::gpu::GPUThreadMappingAttr::get(ctx, mlir::gpu::Threads::DimY);
-  }
-  Attribute threadZ() const {
-    return mlir::gpu::GPUThreadMappingAttr::get(ctx, mlir::gpu::Threads::DimZ);
-  }
-  Attribute warpX() const {
-    return mlir::gpu::GPUWarpMappingAttr::get(ctx, mlir::gpu::Warps::DimX);
-  }
-  Attribute warpY() const {
-    return mlir::gpu::GPUWarpMappingAttr::get(ctx, mlir::gpu::Warps::DimY);
-  }
-  Attribute warpZ() const {
-    return mlir::gpu::GPUWarpMappingAttr::get(ctx, mlir::gpu::Warps::DimZ);
-  }
-  Attribute linearIdX() const {
-    return mlir::gpu::GPULinearIdMappingAttr::get(ctx,
-                                                  mlir::gpu::LinearId::DimX);
-  }
-  Attribute linearIdY() const {
-    return mlir::gpu::GPULinearIdMappingAttr::get(ctx,
-                                                  mlir::gpu::LinearId::DimY);
-  }
-  Attribute linearIdZ() const {
-    return mlir::gpu::GPULinearIdMappingAttr::get(ctx,
-                                                  mlir::gpu::LinearId::DimZ);
-  }
-};
-
-struct MatmulStrategy : StrategyBase {
+class MatmulStrategy : public AbstractGemmLikeStrategy {
+ public:
   MatmulStrategy(MLIRContext *context,
                  const transform_ext::MatchedMatmulCaptures &captures)
-      : StrategyBase(context), captures(captures) {
+      : AbstractGemmLikeStrategy(context), captures(captures) {
     initDefaultValues();
   }
+
+  MatmulStrategy(const MatmulStrategy &) = default;
+  MatmulStrategy &operator=(const MatmulStrategy &) = default;
 
   /// Constructor quantities.
   transform_ext::MatchedMatmulCaptures captures;
 
-  /// Tile sizes for the workgroup / determines grid size for all known
-  /// reduction strategies. The initial values are set by initDefaultValues();
-  SmallVector<int64_t> blockTileSizes;
-  int64_t reductionTileSize;
-  SmallVector<int64_t> numThreads;
-  SmallVector<int64_t> numWarps;
-  bool useAsyncCopies;
-  bool useMmaSync;
-  int64_t pipelineDepth;
-
   void initDefaultValues();
 
-  int64_t m() const {
+  LogicalResult verify() const;
+
+  int64_t m() const override {
     assert(captures.matmulOpSizes.size() == 3 && "need 3 sizes");
     return captures.matmulOpSizes[0];
   }
-  int64_t n() const {
+  int64_t n() const override {
     assert(captures.matmulOpSizes.size() == 3 && "need 3 sizes");
     return captures.matmulOpSizes[1];
   }
-  int64_t k() const {
+  int64_t k() const override {
     assert(captures.matmulOpSizes.size() == 3 && "need 3 sizes");
     return captures.matmulOpSizes[2];
   }
-  int64_t totalNumThreads() const {
-    int64_t res = 1;
-    for (auto v : numThreads) res *= v;
-    return res;
-  }
-  int64_t totalNumWarps() const {
-    int64_t res = 1;
-    for (auto v : numWarps) res *= v;
-    return res;
-  }
 
-  int64_t lhsCopyVectorSize() const {
-    if (k() % 4 == 0) return 4;
-    if (k() % 2 == 0) return 2;
-    return 1;
-  }
-  int64_t rhsCopyVectorSize() const {
-    if (n() % 4 == 0) return 4;
-    if (n() % 2 == 0) return 2;
-    return 1;
-  }
-  int64_t resCopyVectorSize() const { return rhsCopyVectorSize(); }
+  using AbstractGemmLikeStrategy::MappingInfo;
 
-  struct MappingInfo {
-    SmallVector<int64_t> numThreads;
-    // Explicitly computing the tileSizes is only needed until masked
-    // vectorization properly computes the bounds automatically.
-    SmallVector<int64_t> tileSizes;
-    SmallVector<Attribute> threadMapping;
-  };
-  MappingInfo getBlockMapping() const {
+  MappingInfo getBlockMapping() const override {
     return MappingInfo{/*numThreads=*/{},
                        /*tileSizes=*/{blockTileSizes[0], blockTileSizes[1]},
                        /*threadMapping=*/{blockY(), blockX()}};
   }
+
   // LHS copy is of size mxk.
-  MappingInfo lhsCopyMapping() const {
+  MappingInfo lhsCopyMapping() const override {
     assert(reductionTileSize % lhsCopyVectorSize() == 0 &&
            "vector size must divide reductionTileSize");
     int64_t numThreadsK = reductionTileSize / lhsCopyVectorSize();
@@ -157,7 +81,7 @@ struct MatmulStrategy : StrategyBase {
         /*threadMapping=*/{linearIdX(), linearIdY()}};
   }
   // RHS copy is of size kxn.
-  MappingInfo rhsCopyMapping() const {
+  MappingInfo rhsCopyMapping() const override {
     assert(blockTileSizes[1] % rhsCopyVectorSize() == 0 &&
            "vector size must divide blockTileSizes[1]");
     int64_t numThreadsN = blockTileSizes[1] / rhsCopyVectorSize();
@@ -175,7 +99,7 @@ struct MatmulStrategy : StrategyBase {
         /*threadMapping=*/{linearIdY(), linearIdX()}};
   }
   // RES copy is of size mxn.
-  MappingInfo resCopyMapping() const {
+  MappingInfo resCopyMapping() const override {
     assert(blockTileSizes[1] % resCopyVectorSize() == 0 &&
            "vector size must divide n");
     int64_t numThreadsN = blockTileSizes[1] / resCopyVectorSize();
@@ -193,7 +117,7 @@ struct MatmulStrategy : StrategyBase {
         /*threadMapping=*/{linearIdY(), linearIdX()}};
   }
   // COMPUTE is of size mxn.
-  MappingInfo computeMapping() const {
+  MappingInfo computeMapping() const override {
     return MappingInfo{/*numThreads=*/{numWarps[0], numWarps[1]},
                        /*tileSizes=*/{},
                        /*threadMapping=*/{warpY(), warpX()}};
